@@ -8,13 +8,11 @@ Route::prefix('api')
     ->group(function () {
         Route::get('/meta', Guest::class)->name('meta');
 
-        require 'app/auth.php';
+        require __DIR__.'/app/auth.php';
 
         Route::middleware(['api', 'auth'])
             ->group(fn () => Route::get('/sentry', Sentry::class)->name('sentry'));
 
         Route::middleware(['api', 'auth', 'core'])
-            ->group(function () {
-                require 'app/core.php';
-            });
+            ->group(fn () require __DIR__.'/app/core.php');
     });
