@@ -3,13 +3,12 @@
 namespace LaravelEnso\Core\Http\Controllers;
 
 use Illuminate\Routing\Controller;
-use Illuminate\Support\Facades\App;
-use LaravelEnso\Core\Http\Responses\AppState;
+use LaravelEnso\Core\Services\State\Builder;
 
 class Spa extends Controller
 {
     public function __invoke()
     {
-        return App::make(AppState::class);
+        return (new Builder())->handle();
     }
 }
