@@ -3,11 +3,14 @@
 namespace LaravelEnso\Core\Services;
 
 use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\Config;
 
 class Inspiring
 {
     public static function quote(): string
     {
-        return (new Collection(config('enso.inspiring.quotes')))->random();
+        $quotes = Config::get('enso.inspiring.quotes');
+
+        return Collection::wrap($quotes)->random();
     }
 }
